% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todos.R
\name{todos}
\alias{todos}
\title{Todo List}
\usage{
todos(path = ".")
}
\value{
data.frame
}
\description{
This function helps users to summarise the todo-list in the different documents.
}
\author{
Jiaxiang Li
}
